split apart subgraph.py, more
