+++ pdr_backend/lake/test/test_gql_data_factory.py
@@ -21,7 +21,7 @@ def test_update_gql1(
     mock_get_all_contract_ids_by_owner,
     mock_fetch_filtered_predictions,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     del_network_override(monkeypatch)
@@ -29,7 +29,7 @@ def test_update_gql1(
     _test_update_gql(
         mock_fetch_filtered_predictions,
         tmpdir,
-        _sample_daily_predictions,
+        sample_daily_predictions,
         "2023-11-02_0:00",
         "2023-11-04_0:00",
         n_preds=2,
@@ -42,7 +42,7 @@ def test_update_gql2(
     mock_get_all_contract_ids_by_owner,
     mock_fetch_filtered_predictions,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     del_network_override(monkeypatch)
@@ -50,7 +50,7 @@ def test_update_gql2(
     _test_update_gql(
         mock_fetch_filtered_predictions,
         tmpdir,
-        _sample_daily_predictions,
+        sample_daily_predictions,
         "2023-11-02_0:00",
         "2023-11-06_0:00",
         n_preds=4,
@@ -63,7 +63,7 @@ def test_update_gql3(
     mock_get_all_contract_ids_by_owner,
     mock_fetch_filtered_predictions,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     del_network_override(monkeypatch)
@@ -71,7 +71,7 @@ def test_update_gql3(
     _test_update_gql(
         mock_fetch_filtered_predictions,
         tmpdir,
-        _sample_daily_predictions,
+        sample_daily_predictions,
         "2023-11-01_0:00",
         "2023-11-07_0:00",
         n_preds=6,
@@ -84,7 +84,7 @@ def test_update_gql_iteratively(
     mock_get_all_contract_ids_by_owner,
     mock_fetch_filtered_predictions,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     del_network_override(monkeypatch)
@@ -100,7 +100,7 @@ def test_update_gql_iteratively(
         _test_update_gql(
             mock_fetch_filtered_predictions,
             tmpdir,
-            _sample_daily_predictions,
+            sample_daily_predictions,
             st_timestr,
             fin_timestr,
             n_preds=n_preds,
@@ -189,7 +189,7 @@ def test_load_and_verify_schema(
     mock_get_all_contract_ids_by_owner,
     mock_fetch_filtered_predictions,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     del_network_override(monkeypatch)
@@ -201,7 +201,7 @@ def test_load_and_verify_schema(
     _test_update_gql(
         mock_fetch_filtered_predictions,
         tmpdir,
-        _sample_daily_predictions,
+        sample_daily_predictions,
         st_timestr,
         fin_timestr,
         n_preds=5,
@@ -235,7 +235,7 @@ def test_get_gql_dfs_calls(
     mock_update,
     mock_get_all_contract_ids_by_owner,
     tmpdir,
-    _sample_daily_predictions,
+    sample_daily_predictions,
     monkeypatch,
 ):
     """Test core DataFactory functions are being called"""
@@ -264,7 +264,7 @@ def test_get_gql_dfs_calls(
         pdr_predictions_record: pl.DataFrame(
             [
                 x.__dict__
-                for x in _sample_daily_predictions
+                for x in sample_daily_predictions
                 if st_ut_sec <= x.timestamp <= fin_ut_sec
             ]
         ).with_columns([pl.col("timestamp").mul(1000).alias("timestamp")])
