+++ pdr_backend/ppss/web3_pp.py
@@ -9,8 +9,9 @@ from web3 import Web3
 
 from pdr_backend.models.feed import Feed
 from pdr_backend.models.slot import Slot
+from pdr_backend.subgraph.subgraph_pending_slots import get_pending_slots
+from pdr_backend.subgraph.subgraph_feed_contracts import query_feed_contracts
 from pdr_backend.util.strutil import StrMixin
-from pdr_backend.subgraph.core_subgraph import get_pending_slots, query_feed_contracts
 from pdr_backend.util.web3_config import Web3Config
