+++ pdr_backend/util/cli_module.py
@@ -12,7 +12,7 @@ from pdr_backend.predictoor.approach3.predictoor_agent3 import PredictoorAgent3
 from pdr_backend.trader.approach1.trader_agent1 import TraderAgent1
 from pdr_backend.trader.approach2.trader_agent2 import TraderAgent2
 from pdr_backend.trueval.trueval_agent import TruevalAgent
-from pdr_backend.util.check_network import check_network_main
+from pdr_backend.analytics.check_network import check_network_main
 from pdr_backend.util.fund_accounts import fund_accounts_with_OCEAN
 
 from pdr_backend.util.cli_arguments import (
