+++ pdr_backend/util/constants.py
@@ -16,6 +16,7 @@ OCEAN_TOKEN_ADDRS = {
 
 S_PER_MIN = 60
 S_PER_DAY = 86400
+S_PER_WEEK = S_PER_DAY * 7
 
 SUBGRAPH_MAX_TRIES = 5
 WEB3_MAX_TRIES = 5
