+++ pdr_backend/util/constants_opf_addrs.py
@@ -1,14 +1,11 @@
-from enforce_typing import enforce_types
+from typing import Dict
 
-from pdr_backend.util.constants import (
-    SAPPHIRE_TESTNET_CHAINID,
-    SAPPHIRE_MAINNET_CHAINID,
-)
+from enforce_typing import enforce_types
 
 
 @enforce_types
-def get_opf_addresses(chain_id: int):
-    if chain_id == SAPPHIRE_TESTNET_CHAINID:
+def get_opf_addresses(network_name: str) -> Dict[str, str]:
+    if network_name == "sapphire-testnet":
         return {
             "predictoor1": "0xE02A421dFc549336d47eFEE85699Bd0A3Da7D6FF",
             "predictoor2": "0x00C4C993e7B0976d63E7c92D874346C3D0A05C1e",
@@ -18,7 +15,7 @@ def get_opf_addresses(chain_id: int):
             "dfbuyer": "0xeA24C440eC55917fFa030C324535fc49B42c2fD7",
         }
 
-    if chain_id == SAPPHIRE_MAINNET_CHAINID:
+    if network_name == "sapphire-mainnet":
         return {
             "predictoor1": "0x35Afee1168D1e1053298F368488F4eE95E891a6e",
             "predictoor2": "0x1628BeA0Fb859D56Cd2388054c0bA395827e4374",
@@ -65,4 +62,4 @@ def get_opf_addresses(chain_id: int):
             "dfbuyer": "0x2433e002Ed10B5D6a3d8d1e0C5D2083BE9E37f1D",
         }
 
-    raise ValueError(chain_id)
+    raise ValueError(network_name)
