+++ pdr_backend/util/test_noganache/test_topup.py
@@ -1,29 +1,18 @@
 from unittest.mock import MagicMock, patch
 
+from enforce_typing import enforce_types
 import pytest
 
 from pdr_backend.models.token import NativeToken, Token
-from pdr_backend.ppss.ppss import PPSS, fast_test_yaml_str
+from pdr_backend.ppss.ppss import mock_ppss
+from pdr_backend.util.mathutil import to_wei
 from pdr_backend.util.topup import topup_main
-
-
-@pytest.fixture(name="mock_ppss_")
-def mock_ppss(tmpdir):
-    web3_config = MagicMock()
-    web3_config.w3.eth = MagicMock()
-    web3_config.w3.eth.chain_id = 23294
-
-    s = fast_test_yaml_str(tmpdir)
-    ppss = PPSS(yaml_str=s, network="development")
-    ppss.web3_pp = MagicMock()
-    ppss.web3_pp.web3_config = web3_config
-    return ppss
 
 
 @pytest.fixture(name="mock_token_")
 def mock_token():
     token = MagicMock(spec=Token)
-    token.balanceOf.side_effect = [500 * 1e18, 0, 0]  # Mock balance in wei
+    token.balanceOf.side_effect = [to_wei(500), 0, 0]  # Mock balance in wei
     token.transfer.return_value = None
     return token
 
@@ -31,7 +20,7 @@ def mock_token():
 @pytest.fixture(name="mock_native_token_")
 def mock_native_token():
     native_token = MagicMock(spec=NativeToken)
-    native_token.balanceOf.side_effect = [500 * 1e18, 0, 0]  # Mock balance in wei
+    native_token.balanceOf.side_effect = [to_wei(500), 0, 0]  # Mock balance in wei
     native_token.transfer.return_value = None
     return native_token
 
@@ -46,17 +35,17 @@ def mock_get_opf_addresses():
     )
 
 
-def test_topup_main(
-    mock_ppss_, mock_token_, mock_native_token_, mock_get_opf_addresses_
-):
-    with patch("pdr_backend.util.topup.Token", return_value=mock_token_), patch(
-        "pdr_backend.util.topup.NativeToken", return_value=mock_native_token_
-    ), patch(
-        "pdr_backend.util.topup.get_opf_addresses", mock_get_opf_addresses_
-    ), patch(
-        "pdr_backend.util.topup.sys.exit"
+@enforce_types
+def test_topup_main(mock_token_, mock_native_token_, mock_get_opf_addresses_, tmpdir):
+    ppss = mock_ppss("5m", ["binance c BTC/USDT"], "sapphire-mainnet", str(tmpdir))
+
+    PATH = "pdr_backend.util.topup"
+    with patch(f"{PATH}.Token", return_value=mock_token_), patch(
+        f"{PATH}.NativeToken", return_value=mock_native_token_
+    ), patch(f"{PATH}.get_opf_addresses", mock_get_opf_addresses_), patch(
+        f"{PATH}.sys.exit"
     ) as mock_exit:
-        topup_main(mock_ppss_)
+        topup_main(ppss)
 
         mock_exit.assert_called_with(0)
